# cohorts/2026/homework/02-development/submission/openapi.yaml
openapi: 3.0.3
info:
  title: Flowboard API
  version: 1.0.0
paths:
  /api/tasks:
    get:
      summary: List tasks
      responses:
        '200':
          description: Tasks grouped by board state
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Task' }
    post:
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TaskCreate' }
      responses:
        '201':
          description: Created task
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Task' }
  /api/tasks/{task_id}:
    parameters:
      - in: path
        name: task_id
        required: true
        schema: { type: integer }
    patch:
      summary: Update a task
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TaskUpdate' }
      responses:
        '200':
          description: Updated task
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Task' }
        '404': { description: Task not found }
    delete:
      summary: Delete a task
      responses:
        '204': { description: Task deleted }
        '404': { description: Task not found }
components:
  schemas:
    Task:
      type: object
      required: [id, title, description, priority, status, created_at]
      properties:
        id: { type: integer }
        title: { type: string, minLength: 1, maxLength: 120 }
        description: { type: string, maxLength: 500 }
        priority: { type: string, enum: [low, medium, high] }
        status: { type: string, enum: [todo, in_progress, done] }
        created_at: { type: string, format: date-time }
    TaskCreate:
      type: object
      required: [title]
      properties:
        title: { type: string, minLength: 1, maxLength: 120 }
        description: { type: string, maxLength: 500, default: '' }
        priority: { type: string, enum: [low, medium, high], default: medium }
    TaskUpdate:
      type: object
      properties:
        title: { type: string, minLength: 1, maxLength: 120 }
        description: { type: string, maxLength: 500 }
        priority: { type: string, enum: [low, medium, high] }
        status: { type: string, enum: [todo, in_progress, done] }